monashthesisdown: A package for creating undergraduate, Masters, and PhD theses using R Markdown

Using the bookdown package and LaTeX and Word thesis templates from Reed College, this package aims to encourage reproducible research using simple Markdown syntax while embedding all of the R code to produce plots and analyses as well. This package is adapted for Smith College

thesis_gitbook

Creates an R Markdown thesis template as a webpage

thesis_pdf

Creates an R Markdown thesis template as a PDF

thesis_word

Creates an R Markdown thesis template as a Microsoft Word document

thesis_epub

Creates an R Markdown thesis template as an ebook

Author

Maintainer: Benjamin S. Baumer ben.baumer@gmail.com (ORCID)

Authors:

Chester Ismay chester.ismay@gmail.com

Nick Solomon nick@nicksolomon.me